tibbleOne formula handler

tibble_one() accepts formula input. This function translates a formula input value into three types of table one variables (e.g., row, stratification, and by). It is unlikely that you will want to use this function directly, but may wish to use it to debug a formula object that is not providing the tibble one results you were expecting.

parse_tb1_formula(formula, data)

Arguments

formula
a one sided formula object with row variables to the left and stratification variables to the right of a | symbol. For example, formula = ~ x + y | z.

data
a data frame to evaluate formula terms in.

Examples

df = data.frame(a = 1, b = 2, c = 3)

get_tb1_vars(~ a + b | c, df)

#> Error in get_tb1_vars(~a + b | c, df): could not find function "get_tb1_vars"

# identical to

get_tb1_vars(~ . | c, df)

#> Error in get_tb1_vars(~. | c, df): could not find function "get_tb1_vars"

# specify a by variable using *

get_tb1_vars(~ . | b*c, df)

#> Error in get_tb1_vars(~. | b * c, df): could not find function "get_tb1_vars"